+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 
 ### Features
 - Filter data on current bounding box (also known as extent)
+- Show all data of a certain asset type
 - Add a change log